Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

As Jacob says the blade hardware does not support this. Two finger touch is all the controller can support.

Wait to see what comes out of 2914. I would prefer the option to be a setting in CMParts so users can unable it themselves once being made aware that this doesn't work at all properly. But adding it as a default setting... no

Patch-set: 1