Update patch set 1
Patch Set 1:

Yeah, a pref in CMParts is what I was talking about. To be fair, it's not working "not at all" - I'm using it right now to type this. It just has limitations resulting in the x and y coordinates of the two points being swapped sometimeas, as visualized by Multitouch Visualizer. And IIRC "distinct" multitouch just means being able to track two fingers independently. For more than two, there's "jazzhand".

Patch-set: 1