Update patch set 1
Patch Set 1: Abandoned

I just had a look at the internals of the PackageManagerService class and found out that the build system doesn't do anything fancy to the feature xml files anyway (like wrap them up in a signed jar or something). 

Since it's just reading files from /system/etc/permissions and it's therefore trivial to change it locally without rebuilding everything, writing a UI for this has just dropped near the bottom of my priority list - from "fixes a major hassle" to "would save a minute every once in a while". If I ever do get around to it, I'll of course contribute it.

Patch-set: 1