Update patch set 1
Patch Set 1:

One more thing, I'm adding koush because this will require an update to Blade recovery in ROM Manager.

Also this patch does break compatibility with devices that shipped with 2.1 unless they first update to the new radio. I'm anticipating some user complaints but it's impossible to downgrade so this is the only option.

Patch-set: 1