Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

The testing went very well but I've been delaying responding because I have another kernel change to submit. 

I'm happy for this to be merged as is, I'll add my changes as part of the new sensor lib when its ready.

Patch-set: 1