Update patch set 1
Patch Set 1:

I stand corrected: Signalling (ringtone etc.) in the built-in SIP client is routed okay, but once the called person goes offhook, for some reason the SIP client does a setMode(NORMAL), and audio gets routed to the speaker again :-( . This smells a bit like a bug in the SIP client.

I guess STREAM_VOICE_CALL support is the thing we need. It would probably be possible to check whether a VOICE_CALL stream is open in AudioFlinger and pass a flag through setParameters() down to libaudio. This feels like quite a hack though.

Patch-set: 1