Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This one is a bit strange. With the current setting it defaults to using in-call touch UI, but shows an option in call settings (Force in-call UI) which is enabled by default. With the patch applied in-call touch UI is still used, but no option is shown in call settings.

Any way, this patch looks right.

Patch-set: 1