Update patch set 1
Patch Set 1:

Maybe I've changed something locally in the source causing the force touch UI to default to on, but I can't find it. The other device I work on (U8220) has no proximity sensor and is working as it should with touch UI disabled (ie. its not defaulting to force touch UI).

Patch-set: 1