Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

There were 2 causes for the radio not turning off some times (or hiss after radio turned off). The radio jni was failing to turn off the radio in some situations, and doRoutingOrMute was not being triggered on libaudio to mute the output. This should fix the muting.

Patch-set: 3