Blade: Small config app to enable/disable USB-charging (stock ZTE ROMs have this option)
I wrote a configuration app for the ZTE V9 port to handle touchscreen calibration and some
other things and I added the USB-charging option to it as it is something that stock ZTE
ROMs usually have as an option.

Since the work has allready been done I figured the Blade could have use for a version of
the app with all the calibration stuff removed (which leaves only the USB-charging option).

Not incredible useful but some users might want to have the option.

Change-Id: I71da3c9a7784db7c5d9f4cd91af23343c51b830d