Update patch set 1
Patch Set 1:

Been testing for a couple days. Seems stable, but twice it showed me the "Shutting down..." dialog and turned off when the battery was still a little less than 20%. Also, the Tetris app is no longer opening, which I'm assuming is due to this patch but could be something else.

Patch-set: 1