Update patch set 3
Patch Set 3:

The shutdowns sound like the high temp bug, but that shouldn't be happening. Try to get some logs (both logcat & dmesg). Some blades including the 2 I have don't have that problem, so I can't replicate it.

Its normal for it not to get hot, it doesn't really overheat, it gets incorrect readings for a while after resume. This also happens in ZTE roms, ZTE have just disabled the code in the framework that shuts down the device if it overheats (not good if it really overheats).

Patch-set: 3