Update patch set 1
Patch Set 1:

I would rather have it in settings. Moving it to cmparts will make it harder to find and more likely for users to miss. The proximity sensor calibration function in this app needs to be easy to find since a lot of users now have proximity sensor problems with the default calibration.

Patch-set: 1