Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

If proprietary files are pulled from the device or an update.zip, lights.msm7k.so will not be present in vendor/zte/blade/proprietary and build will fail. The PRODUCT_COPY_FILES entry for lights.msm7k.so should be removed from unzip- and extract-files.sh to make it build again.
Other than that, looks good to me.

Patch-set: 3