Update patch set 7
Patch Set 7: Looks good to me, but someone else must approve

I've been using this patch7 for about a week now and see no problems, but people using patch4 are reporting battery issues.

So can we get this merged so we can sort out if these are related to the LED code or just people with dodgy settings/batteries?

Patch-set: 7