Update patch set 3
Patch Set 3: No score

Sorry to ask this here, but can I insert a new patch set in this "thread" with updated translations for pt-rPT? And how do I add a new patch to this one (a link to an explanation page would be sufficient..).

My knowledge of git/repo is minimal, and I can't seem to find how to update patches, only how to create new ones.

Again, sorry for polluting this thread here.

Patch-set: 3