Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

This should be dependent on http://review.cyanogenmod.com/8647   (new strings are added in 8647 which this includes translations for). This will cause a compile error if merged before 8647. Do not merge yet.

Otherwise, looks good to me. Once 8647 is merged it will be fine.

Patch-set: 3