Update patch set 3
Patch Set 3:

Simon - Camera shutdown will not be effected by any of these changes. If you are talking about the old sensors any problems with them would have been there in previous kernels. If you are specifically talking about the new sensors then all this patch does is enable the code that was already there. It is difficult for me to test anything with the new hardware since I dont have physical access to it.

Patch-set: 3