Update patch set 3
Patch Set 3:

Auto brightness is a bit confusing atm. In a dim light or almost dark I get only a -1 when I used to get some reading (1..20). This causes the screen to stay at whatever the brightness was when the screen was on the last time. Partly covering the sensor in bright light and then exposing it to full light again causes the sensor reading to stay at the lower level. Also, I can't get "Reset threshold" under "Light sensor filter" to work at all. Maybe it's because the values are in lux (min.400) and the new sensor values are in the range of 0..255? If we are going to keep it like this, the defaut AB settings should also be updated, yes?

Patch-set: 3