Update patch set 3
Patch Set 3:

The light sensor values were already screwed up. The change to the 2.6.35 kernel changed the light sensor setup and the thresholds have not been adjusted for that, but this change should be putting them back to where they were. I still get good values with these settings (as I did with the previous 2.6.35 settings that were also causing complaints) but it looks like there are no values that work on all hardware. I give up. If it doesnt work for you dont use it.

Patch-set: 3