Update patch set 3
Patch Set 3:

The new default prox settings work fine for me but I have to agree the light sensor is messed up.

Now it's returning very low numbers.
Under 'edit other levels' in the previous kernel it was returning 200-300 for the indoor lighting (using a bright daylight bulb) I'm currently in.  With this new patchset it returns 2 to 6.

There are lots of reports of this at Modaco, whereas I hadn't seen any complaints about the Froyo to GB change?  

Now I myself have been posting that different doesn't mean wrong and you just need to change the levels to match the new sensor readout.  But based on Sakiris comments above, I've had a play and there isn't enough granularity between the values.  If I just position my body between the phone and the light so it's in shadow, it records 0 which should be pitch black but is in fact still bright enough that if I had a TFT I might want the brightness above min.  The previous kernel reports 50.

Personally I can live without auto brightness. I have my edit other levels to stick the screen on min indoors and max outdoors. I can achieve that with a toggle button on the home screen rather than auto-brightness!  I'm just backing up the comment that on at least some hardware (I have an early UK gen 1 oled) this change is doing the opposite of your aim of restoring previous behaviour.  In that the sensor return values were previously usable in the GB kernel but now aren't.

Now I know one of the problems (and maybe the one we're probably trying to get around here?) is an inability setup defaults useful for all variations of the Blade.  But here, it's not just that the defaults are wrong.  The figures being returned are such that they can't possibly be set right, even by the end user.  Never mind in a CM supplied default.

Patch-set: 3