Update patch set 4
Patch Set 4:

On the previous GB kernel, the default calibration didn't work for me but the calibrate button worked. Giving me values of 350 - 450-ish.  It only worked indoors though, in bright outdoor light it failed to calibrate.  I suspect the time of year is going to prevent me from testing this in the sun again for the next few months!

The new defaults of 8000-8500 do work for me though.  And if I choose to calibrate I get a score of 6738-8422 which also works.

The light values are now returned to normal.

As for calibration issues, a poster here had trouble with a build based on patchset 3: http://android.modaco.com/topic/335328-cyanogenmod-7-android-237-is-now-gen-2-only/page__view__findpost__p__1846446
I don't know if he previously had issues with the nightlies or not.

Patch-set: 4