Update patch set 4
Patch Set 4:

That header was straight from the ZTE kernel source, and as you would know, ZTE love white space errors. I wanted to leave it unchanged, but since I've already had to modify it slightly it wont hurt to fix the whitespace errors as well (btw, your file is full of tabs and has an error on L24).

Patch-set: 4