Update patch set 7
Patch Set 7:

Tom, wbaw :

I think the wifi issue happens but not related with wipe method or modules either. Most of them are getting disconnect issue as found with this logcat : 

http://pastebin.com/sCZS8dq3

Maybe we need to dig in deeper, asking whether they have tpt-upgraded and which version they are using indeed.

burst.

Patch-set: 7