Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

Seems to work OK, added 10mins to the build time, but that included downloading, so expect about 5mins (10%) usually.

So are we always building the kernel from source now, I thought it was a case of if there was a prebuilt, then use it?

I guess it does help keep things in sync and easier than merging stuff in from github manually.

Patch-set: 3