blade: Fix video recording (part 1/2)
This should fix video recording on the Blade.

The changes to the camera HAL are mainly cleanups and edits from C3C0. The work getting video recording was done by me.

This is the first poart of the commit. The second part is at the propietary side and a pull request can be found at https://github.com/koush/proprietary_vendor_zte/pull/1 which fixes this issue and adds the ZTE Skate as well.

Change-Id: I309c6d104d00ab0429df9ff1852338cc89b9ab85