Update patch set 8
Patch Set 8:

Yes. There is a problem playing back the videos on the phone. External apps work fine but natively it doesn't work

It is a code issue. The software decider tries to read metadata and plays it in frame by frame mode and then crashes on one or both of them.

Patch-set: 8