Update patch set 2
Patch Set 2:

I'm inclined to agree with the assessment that this is a good idea.  On a phone with such a small rom, all that included media is pretty wasteful.  Especially the 10mb of sample videos.

I understand this patch fixes the boot animation backup: http://review.cyanogenmod.com/#/c/14763/
So it's hard to see any downside of this combination of patches?

Patch-set: 2