Update patch set 1
Patch Set 1:

And we can't even play the HD videos which are responsible for most of the space difference.
I'm kind of with you in that it doesn't make that much difference as we can make smaller builds easily ourselves.  Heck, even if you're not compiling it yourself and using an official build it's simple to open up the zip and just delete the videos folder.

I think the point here is the sheer uselessness of the full version.  There appears to be no downside to going mini?

Patch-set: 1