Update patch set 2
Patch Set 2:

It does look mostly to just be re-arranged yeah.
I've sat and compared them by hand as best I can (and it compiles and boots)

As far as I can tell Boardconfig.mk is identical

device.mk is missing the following from the new version:
libOmxVdec
libOmxVenc
and line 100 in the old version doesn't appear to be represented in the new version.
The rest of it appears the same.

So if those 3 differences in device.mk are intentional rather than overlooked then I guess this is good?

Patch-set: 2