Update patch set 9
Patch Set 9:

Switched to AAC audio. I had choppy/pitched audio playback with AMR-NB possibly because of incorrect sample rate. AAC works fine and it's higher quality anyway.

As far as I can tell, issue with the camcorder video playback is the same we have with other "high" quality video (missing OMX). No reason to hold back this patch because of that. If we wanted to play the recorded videos, we would have to somehow cap the camcorder to record only at 320x240. Not much point in that. ;)

With this patch and the needed proprietary lib, video recording works ok. Video (and audio) playback is fine on computer and on device with third party app (e.g. VLC media player).

Patch-set: 9