Update patch set 9
Patch Set 9: I would prefer that you didn't submit this

The issue with playback is actually quite different to with other media. Working OMX codecs would not fix it (although I guess they would since the disablemeta flag shouldn't be needed). The problem with playback is that stagefreight appears to not support video that doesn't include meta data, but to get recording to work we need to disable meta data.

I've also seen some other bugs related to video recording that I'd like to look into further. I haven't had time over the last week to look at this and if I don't have time in the next few days I will merge it as it is, but I would prefer to play with it more before merging.

Patch-set: 9