Update patch set 9
Patch Set 9: Looks good to me, but someone else must approve

i noticed a lot of other devices seem to be setting debug.camcorder.disablemeta=1 which seems to be the only way to get recording working without OMX, but does break playback on the phone due to missing metadata that is being looked for. 

to a degree that sounds more like a bug in the video player, as metadata shouldn't be mandatory.

Patch-set: 9