Update patch set 1
Patch Set 1: Do not submit

My view is this should not be merged, the phone has space for the full size build, so we should use it, but I have made some progress on a solution that I think should make everyone happy. It will hopefully be finished soon & I believe will be more versatile than this, allowing other things that are not essential (like VideoEditor) to be left out for smaller /system partitions.

Patch-set: 1