Update patch set 3
Patch Set 3:

I'm not removing support for live wallpapers (and I doubt thats what you really meant). The live wallpapers themselves are fairly small so not really worth removing.

msm_mdp.h should match the header in the kernel, we shouldn't be adding code to it that isn't in the kernel. That is the main reason I applied all those 3.0 patches. I'll apply the recent patches to the kernel and update the headers when I have some time (hopefully in the next few days).

Patch-set: 3