Update patch set 2
Patch Set 2:

Hi Tom, if you believe this is correct, I think you can submit. I still haven't tested with the permissions "system" but if they give rw permissions as with "root", then I don't see why it would fail. 

Btw, this patch is just a side-product of another attempt to try to nail down something in the sensors' drivers. I started to think that the code is using the compass to simulate the Gsensor, but now I'm not so sure. The gsensor seems to draw lots of power when it is on..

Thanks

Patch-set: 2